:bug: Fixed bug when img string is empty

+++ routes/routes.js
@@ -380,19 +380,30 @@ function routes() {
 					});
 				} else {
 					userSchema.findOne({ _id: user_id }, async (err, doc) => {
+
 						if (err) throw err;
 						let oldimg = doc.img;
-						fs.unlink('public/uploads/'+oldimg, (err) => {
-							if (err) throw err;
-						  });
-						doc.img = req.file.filename;
-						await doc.save();
 
-						console.log(req.file.filename);
-						res.redirect("/profile", 200, {
-							msg: "File uploaded",
+
+						if (oldimg == ""){
+							doc.img = req.file.filename;
+							await doc.save();
+							console.log('Toegevoegd als nieuwe PF');
+						}
+						else{
+							fs.unlink('public/uploads/'+oldimg, (err) => {
+								if (err) throw err;
+							  });
+							  doc.img = req.file.filename;
+							  await doc.save();
+							  console.log('vervangen');
+						}
+						res.redirect('/profile', 200, {
+							msg: 'File uploaded',
+              
 							file: `uploads/${req.file.filename}`
 						});
+
 					});
 				}
 			}
